agent-ui-capture-intelligent/requirements.txt
playwright>=1.48.0
pydantic>=2.9.0
python-dotenv>=1.0.1
openai>=1.51.0
pytest-asyncio>=0.24.0
